+++ test/CMakeLists.txt
@@ -1,4 +1,4 @@
-file(GLOB TEST_SRC_FILES ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
+file(GLOB_RECURSE TEST_SRC_FILES ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
 
 find_package(OpenMP REQUIRED)
